feat: Application rename
