+++ pom.xml
@@ -9,11 +9,11 @@
 		<version>3.3.2</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.luv2code</groupId>
-	<artifactId>cruddemo</artifactId>
+	<groupId>com.rvo</groupId>
+	<artifactId>school-crud-api</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>cruddemo</name>
-	<description>Demo project for Spring Boot</description>
+	<name>schoolcrudapi</name>
+	<description>School CRUD project for Spring Boot</description>
 	<url />
 	<licenses>
 		<license />
