--- src/main/java/com/luv2code/cruddemo/enums/Role.java
@@ -1,5 +1,0 @@
-package com.luv2code.cruddemo.enums;
-
-public enum Role {
-    TEACHER, PRINCIPAL, ADMIN
-}
