--- src/main/java/com/luv2code/cruddemo/exception/StudentErrorResponse.java
@@ -1,15 +1,0 @@
-package com.luv2code.cruddemo.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class StudentErrorResponse {
-    private int status;
-    private String message;
-    private long timeStamp;
-
-}
