--- src/main/java/com/luv2code/cruddemo/exception/StudentNotFoundException.java
@@ -1,17 +1,0 @@
-package com.luv2code.cruddemo.exception;
-
-public class StudentNotFoundException extends RuntimeException {
-
-    public StudentNotFoundException(String message) {
-        super(message);
-    }
-
-    public StudentNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public StudentNotFoundException(Throwable cause) {
-        super(cause);
-    }
-
-}
