+++ src/main/java/com/rvo/schoolcrudapi/dto/StudentUpdateRequest.java
@@ -1,0 +1,13 @@
+package com.rvo.schoolcrudapi.dto;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+import lombok.Data;
+
+@Data
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class StudentUpdateRequest {
+    private String firstName;
+    private String lastName;
+    private String email;
+}
