+++ src/main/java/com/rvo/schoolcrudapi/dto/TeacherDummyDataPayload.java
@@ -1,0 +1,20 @@
+package com.rvo.schoolcrudapi.dto;
+
+import java.util.List;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class TeacherDummyDataPayload {
+
+    @JsonProperty("name")
+    private String fullName;
+    private String email;
+    private List<String> subjects;
+}
