+++ src/main/java/com/rvo/schoolcrudapi/enums/Role.java
@@ -1,0 +1,5 @@
+package com.rvo.schoolcrudapi.enums;
+
+public enum Role {
+    TEACHER, PRINCIPAL, ADMIN
+}
