+++ src/main/java/com/rvo/schoolcrudapi/exception/GlobalExceptionHandler.java
@@ -1,0 +1,39 @@
+package com.rvo.schoolcrudapi.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+import com.fasterxml.jackson.databind.JsonMappingException;
+
+@ControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<StudentErrorResponse> handleGenericException(Exception ex) {
+
+        StudentErrorResponse error = new StudentErrorResponse();
+        error.setStatus(HttpStatus.BAD_REQUEST.value());
+        error.setMessage(ex.getMessage());
+        error.setTimeStamp(System.currentTimeMillis());
+
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(error);
+    }
+
+    // @ExceptionHandler(Exception.class)
+    // public ResponseEntity<String> handleGenericException(Exception ex) {
+    // return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An
+    // unexpected error occurred.");
+    // }
+
+    @ExceptionHandler(JsonMappingException.class)
+    public ResponseEntity<StudentErrorResponse> handleJsonMappingException(JsonMappingException ex) {
+        StudentErrorResponse error = new StudentErrorResponse();
+        error.setStatus(HttpStatus.BAD_REQUEST.value());
+        error.setMessage("Invalid JSON: " + ex.getMessage());
+        error.setTimeStamp(System.currentTimeMillis());
+
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(error);
+    }
+}
