+++ src/main/java/com/rvo/schoolcrudapi/exception/StudentErrorResponse.java
@@ -1,0 +1,15 @@
+package com.rvo.schoolcrudapi.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class StudentErrorResponse {
+    private int status;
+    private String message;
+    private long timeStamp;
+
+}
