+++ src/main/java/com/rvo/schoolcrudapi/exception/StudentNotFoundException.java
@@ -1,0 +1,17 @@
+package com.rvo.schoolcrudapi.exception;
+
+public class StudentNotFoundException extends RuntimeException {
+
+    public StudentNotFoundException(String message) {
+        super(message);
+    }
+
+    public StudentNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public StudentNotFoundException(Throwable cause) {
+        super(cause);
+    }
+
+}
