+++ src/main/java/com/rvo/schoolcrudapi/model/Teacher.java
@@ -1,0 +1,44 @@
+package com.rvo.schoolcrudapi.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "teacher")
+@Data
+@NoArgsConstructor
+public class Teacher {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    @Setter(AccessLevel.NONE)
+    private Integer id;
+
+    @Column(name = "first_name")
+    private String firstName;
+
+    @Column(name = "last_name")
+    private String lastName;
+
+    @Column(name = "email")
+    private String email;
+
+    @Column(name = "subject")
+    private String subject;
+
+    public Teacher(String firstName, String lastName, String email, String subject) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.email = email;
+        this.subject = subject;
+    }
+}
