+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 # # Turn off the Spring Boot Banner
 # spring.main.banner-mode=off
 
-# spring.application.name=cruddemo
+# spring.application.name=school_crud_api
 
 # spring.datasource.url=jdbc:mysql://localhost:3306/student_tracker
 # spring.datasource.username=springstudent
