+++ src/main/resources/application.yaml
@@ -2,7 +2,7 @@ spring:
   main:
     banner-mode: off
   application:
-    name: cruddemo
+    name: school_crud_api
   datasource:
     url: jdbc:mysql://localhost:3306/student_tracker
     username: springstudent
