--- src/test/java/com/luv2code/cruddemo/CruddemoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.luv2code.cruddemo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CruddemoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
