+++ src/test/java/com/rvo/schoolcrudapi/SchoolCrudApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.rvo.schoolcrudapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SchoolCrudApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
